+++ src/main/java/com/example/spinlog/global/security/oauth2/handler/login/OAuth2LoginSuccessHandler.java
@@ -1,6 +1,5 @@
-package com.example.spinlog.global.config.oauth2.handler;
+package com.example.spinlog.global.security.oauth2.handler.login;
 
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
