+++ src/main/java/com/example/spinlog/global/security/oauth2/user/OAuth2Response.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.user.security.dto;
+package com.example.spinlog.global.security.oauth2.user;
 
 public interface OAuth2Response {
